Fix python::pip installing $editable VCS packages every puppet run
This is due to VCS packages appearing in `pip freeze --all` as:
-e git://example.com/package.git@<ref>#egg=<egg>

Updates $grep_regex, which checks whether a given version is installed, to
match against the two different output formats of `pip list`. No longer
relies on pip freeze.

Additionally, this commit:

- Removes the multiple exec resources which were mostly the same and replaces
with variables $pip_exec_name, $command, and $unless_command, which get
used in a single exec resource at the end of the manifest.

- Checks if $pkgname contains ==, and if so splits based on == and sets
$real_pkgname to the first part, and $_ensure to the second part. Uses of
$pkgname have been replaced with $real_pkgname.

Fixes #193